Separate file for pysb-petab
