+++ python/sdist/amici/petab_import_pysb.py
@@ -1,0 +1,1 @@
+../../amici/petab_import_pysb.py
